Fixed EI, DI, etc.
Corrected issue where CPU was not properly returning to the stored PC in the stack after an interrupt request. The issue was traced to incorrect push method (SP not being decremented before storing PC).

Fixed miscellaneous opcodes as well, such as AND, A; CB $87, and CB $C7.

Begun to test timer as well. Seems to not be working according to Blargg's test rom.